refactor thor actios extend to module
