+++ lib/dockerize_stack.rb
@@ -1,10 +1,10 @@
 require 'thor'
-require_relative './dockerize_stack/thor_extend.rb'
+require_relative './thor_extend.rb'
 require_relative './dockerize_stack/react.rb'
 require_relative './dockerize_stack/rails.rb'
 
 module DockerizeStack
-  class Command < DockerizeStack::ThorExtend
+  class Command < Thor
     include Thor::Actions
 
     option :path
